Restructure container divs to capture nums individually
Wrapped nums buttons inside of a div container with class nums,
and wrapped nums container and the selector buttons (clear, +/-
and =) together in a div container with class btns. The reason the
selector buttons are in the same group as the nums it to position them
beneath the nums on the screen.