Add event listeners to the buttons
Added event listeners to the num buttons, the clear button,
and the = button. At the moment the nums event listener prints
the number selection to the result div (i.e. screen) and adds a space.

The clear button clears the text display on the result div (i.e. screen)

The = button calls the operate() and displayResult() functions which
run and display the result in the result div (i.e. screen).