Update num and operator event handler functionality
Changed the functionality of the numhandler on the number buttons to
display a num and then display subsequent clicks of numbers
consecutively without a space.

Added functionality to the operator handler to add a class
to the selected operate to change it's background color dynamically by
adding an active class and removing that active class from all other
sibling operators.