Add functions to operate calculator
Added functionality to read user input (i.e. read the number that the
user clicks into the calculator using the number buttons), read the
user-selected operator, and then read a second number in order to
call the operate function and subsequently call one of the operational
functions based on the user-selected operator, and the user selected
numbers.

The first number and operator are stored in an array called
storageValues which updates itself each time its length === 3, that is,
each time that the array contains two numbers and an operator, the
operate function is called, the array is wiped, and the result of the
operate function plus the subsequent operator to be used are stored in
preparation for the user to select another number.