Add divs and classes to index.html
Add clear button, percentage sign, and +/- change sign button and an
equal button. Add classes to these buttons in order to set an event
listener in JS. Add divs to display a user message with the current
value of the operation, an operation message showing the current
operation, and an error message showing any error message to the user
about the usage of the calculator.