Add functions to program a digital calculator
Added functions for operations, namely: add, subtract, multiply, divide.
Added an operate function to read the user-selected operator and
call the correct operation function. Added event listeners to the
numbers that update the screen's class list with a 'number-selected'
class whenever a user selects a number. The same happens when a user
selects an operator, and the equal button, respectively. These classes
serve as flags to the program so that values can be stored properly and
so that event listeners can be toggled off in the necessary situations,
e.g., if the user selects 2 + 2 =, the screen will display 4. The screen
will also have a result class (serving as a flag) to indicate to the
numberHandler function that if the user selects a number, the number
should print to a blank screen as it is a new and unique number, not
get appended to the result that is presently displayed on the screen.

Additionally there are handler functions for the clear, +/-, =, and
all operator buttons. Each function, together with a system of adding
and removing flags from the screen class list, works together to achieve
the accurate operation of the calculator.